Don't have a giant singleton
